fix: element goes to display none when toggling device toolbar in the devtools
